Optimise the Card->Reader Manchester decoder in Sniffing to make it faster
It now decode and sample in the same time, so the decode half bit sampling to Manchester in the CodecTask can be removed.

This optimisation may also fix the bug that the CodecTask not finished after start of Reader->Card Sampling and make the reader->Card sampling not in perfect timing and cause the  Reader ->Card Sniffed data to be wrong.

Trace from oscilloscope for the Problem:

ReaderSniffing Enable
    __________                                    _________
__|                |____________________|              |_________
CardSniffing Enable
                            _____________                          _______
________________|                     |______________|
Task
                       __                           ___----
_|__|__|__|___|  |__|__|__|__|__|_|           |__|__|__|__|

"-" indicate the problem

(cherry picked from commit a246b98)